WIP: Break up: Don't use mod.rs and don't set module name to same as dependency
